requirements.txt
pywebview>=4.0
